move game code to game.go
